refactore code to store data in postgre infrastructure, to prevent accidental database flushes

+++ src/data/store_jps.py
@@ -4,10 +4,6 @@ project_dir = os.path.join(os.path.dirname(__file__), os.pardir, os.pardir)
 sys.path.append(project_dir)
 
 import dotenv
-import geopandas as gpd
-import pandas as pd
-import math
-from timeit import default_timer as timer
 
 from sqlalchemy import create_engine
 from sqlalchemy.engine.url import URL
@@ -15,7 +11,7 @@ from sqlalchemy import MetaData
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.types import TIMESTAMP as typeTIMESTAMP
 
-from src.data.processing_func import (extract_geo_jams, extract_geo_sections)
+from src.data.processing_func import (extract_geo_jams, extract_geo_sections, store_jps)
 
 project_dir = os.path.join(os.path.dirname(__file__), os.pardir, os.pardir)
 dotenv_path = os.path.join(project_dir, '.env')
@@ -38,23 +34,15 @@ meta.bind = engine
 meta.reflect()
 
 #Flush JamPersection and Build geo_sections
-jps = meta.tables["JamPerSection"]
-jps.delete().execute()
-geo_sections = extract_geo_sections(meta)
+flush = None
+while (flush != "Y") and (flush != "N") and (flush != "y") and (flush != "n"):
+    flush = input("Are you sure you want to flush JamPerSection table? (Y/N): ")
+    if (flush == "Y") or (flush == "y"):
+        print("Flushing JamPerSection...")
+        jps = meta.tables["JamPerSection"]
+        jps.delete().execute()
+    elif (flush == "N") or (flush == "n"):
+        print("Aborting operation...")
+        sys.exit()
 
-##Divide the in batches
-batch_size = 20000 #arbitrary number to make the code tractable
-total_rows, = meta.tables["Jam"].count().execute().first()
-number_batches = math.ceil(total_rows / batch_size)
-
-for i in range(0, number_batches):
-    start = timer()
-    #Build and store JamPerSection
-    geo_jams = extract_geo_jams(meta, skip=i*batch_size, limit=batch_size)
-    jams_per_section = gpd.sjoin(geo_jams, geo_sections, how="inner", op="contains")
-    jams_per_section = jams_per_section[["JamDateStart", "JamUuid", "SctnId"]]  
-    jams_per_section["JamDateStart"] = jams_per_section["JamDateStart"].astype(pd.Timestamp)
-    jams_per_section.to_sql("JamPerSection", con=meta.bind, if_exists="append", index=False)
-    end = timer()
-    duration = str(round(end - start))
-    print("Batch " + str(i+1) + " of " + str(number_batches) + " took " + duration + " s to be successfully stored.")
+store_jps(meta)
